research/scanner.py
import whois
from datetime import datetime
import socket
from urllib.parse import urlparse

def get_domain_info(url):
    # Normalize domain extraction
    parsed = urlparse(url if "://" in url else f"http://{url}")
    domain = parsed.netloc.replace("www.", "")
    
    info = {
        "domain": domain, 
        "age_days": -1, 
        "is_ssl": False, 
        "registrar": "Unknown",
        "ip_address": "Unknown"
    }

    # 1. IP Address Lookup (Usually always works)
    try:
        info["ip_address"] = socket.gethostbyname(domain)
    except:
        pass

    # 2. WHOIS & Age Check
    try:
        w = whois.whois(domain)
        created = w.creation_date
        
        # Handle cases where multiple dates are returned
        if isinstance(created, list):
            created = created[0]
        
        if created and isinstance(created, datetime):
            info["age_days"] = (datetime.now() - created).days
            info["registrar"] = w.registrar or "Unknown"
        
        # --- SUPREME OVERRIDE FOR TRUSTED CORPORATIONS ---
        # Big companies (Google, YouTube, Microsoft) use specific registrars.
        # Scammers almost NEVER use these high-security registrars.
        reg_lower = str(w.registrar).lower()
        trusted_corps = ['markmonitor', 'google', 'microsoft', 'csc corporate', 'amazon', 'apple']
        
        if any(corp in reg_lower for corp in trusted_corps):
            if info["age_days"] < 365: # If lookup failed or shows -1
                info["age_days"] = 5000 # Force "Trusted Old" status
                info["registrar"] = f"{w.registrar} (Verified Corporate)"
    except:
        # If WHOIS is totally blocked, we keep age_days at -1
        pass

    # 3. SSL Check
    try:
        socket.create_connection((domain, 443), timeout=1)
        info["is_ssl"] = True
    except:
        pass

    return info