research/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
xgboost
pandas
numpy
python-whois
pyOpenSSL
